+++ app/__generated__/[legal].[disclaimer]._index.server.tsx
@@ -1,0 +1,47 @@
+/* eslint-disable */
+      /* This is a auto generated file for building the project */ 
+
+
+      import type { PageMeta } from "@webstudio-is/sdk";
+      import type { System, ResourceRequest } from "@webstudio-is/sdk";
+export const getResources = (_props: { system: System }) => {
+  const _data = new Map<string, ResourceRequest>([
+  ])
+  const _action = new Map<string, ResourceRequest>([
+  ])
+  return { data: _data, action: _action }
+}
+
+
+      export const getPageMeta = ({
+  system,
+  resources,
+}: {
+  system: System;
+  resources: Record<string, any>;
+}): PageMeta => {
+  return {
+    title: "[DISCLAIMERS] - VAIIYA: Security just got better.",
+    description: "[Disclaimers for the site.]\n\nThink you're safe? Think again. With VAIIYA's impenetrable defenses, You can rest easy. Knowing you are safe from every angle, and evey type of danger.",
+    excludePageFromSearch: true,
+    language: "",
+    socialImageAssetName: undefined,
+    socialImageUrl: "",
+    status: undefined,
+    redirect: "",
+    custom: [
+    ],
+  };
+};
+
+
+      type Params = Record<string, string | undefined>;
+export const getRemixParams = ({ ...params }: Params): Params => {
+  return params
+}
+
+
+      export const projectId = "194e2b25-97be-443f-8441-f7922f347c91";
+
+      export const contactEmail = "chaircomputeroverdrive@gmail.com";
+