+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # CHANGELOG FILE
 ---
 
+
+-- [APR 3 137pm PST]
+
+- ADDED legal/disclaimers/ with GNU GPLv3 notice, copyright, and Embarks stuff and The VAIIYA website project being protected by fair dealing.
+
 -- [APR 3 9:43am PST]
 
 - DELETED BRANCH `WEBSITE_V3_SOURCE`